Add PageViewWithIndicators.
